+++ causalnerve/memory/__init__.py
@@ -2,7 +2,7 @@ from .structural_memory_bank import StructuralMemoryBank
 from .episodic_memory import EpisodicMemory
 from .motif_archive import MotifArchive
 from .recurrence_engine import RecurrenceEngine
-from .replay_engine import StructuralReplayEngine, GraphSnapshot, RevisionRecord
+from .replay_engine import StructuralReplayEngine, GraphSnapshot, RevisionRecord, GraphDiff
 
 __all__ = [
     "StructuralMemoryBank",
@@ -11,5 +11,6 @@ __all__ = [
     "RecurrenceEngine",
     "StructuralReplayEngine",
     "GraphSnapshot",
-    "RevisionRecord"
+    "RevisionRecord",
+    "GraphDiff"
 ]